+++ lib/sunday_ui.dart
@@ -2,7 +2,7 @@ library sunday_ui;
 
 import 'dart:ui';
 
-import 'package:sunday_core/GetGtorage/get_storage.dart';
+import 'package:sunday_get_storage/sunday_get_storage.dart';
 import 'package:sunday_ui/Scripts/Colors/storing_color/storing_color.dart';
 
 /// export specific files
@@ -15,6 +15,11 @@ export 'MainComponents/sunday_list_view/sunday_list_view.dart';
 export 'MainComponents/sunday_text/sunday_text.dart';
 export 'MainComponents/sunday_text_button/sunday_text_button.dart';
 export 'MainComponents/sunday_text_field/sunday_text_field.dart';
+export 'MainComponents/sunday_layout/sidebar_layout/group.dart';
+export 'MainComponents/sunday_layout/sidebar_layout/item.dart';
+export 'MainComponents/sunday_layout/sidebar_layout/view.dart';
+export 'MainComponents/sunday_layout/sunday_layout.dart';
+export 'MainComponents/sunday_layout/sidebar_layout/toogle_button.dart';
 
 /// export scripts
 export 'Scripts/Colors/storing_color/storing_color.dart';
@@ -25,7 +30,6 @@ export 'CoreComponents/sunday_appbar/sunday_appbar.dart';
 export 'CoreComponents/sunday_bottom_bar/sunday_bottom_bar.dart';
 export 'CoreComponents/sunday_navigationbar_item/sunday_navigationbar_item.dart';
 export 'CoreComponents/sunday_scaffold/sunday_scaffold.dart';
-
 
 /// Initializes the UI settings and stores them in the SundayGetStorage.
 ///
@@ -48,7 +52,7 @@ export 'CoreComponents/sunday_scaffold/sunday_scaffold.dart';
 /// Returns:
 ///   Future<void>: A future that completes when the operation is done.
 Future<void> initUI({Object? globals}) async {
-  final box = SundayGetStorage();
+  final box = GetStorage();
 
   globals ??= {
     "default-ui": {